% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fix_data_type.R
\docType{data}
\name{analysis_by_group}
\alias{analysis_by_group}
\title{a}
\format{
Each is a tibble with two variables and three observations
}
\usage{
analysis_by_group
}
\description{
Band membership
}
\details{
These data sets describe band members of the Beatles and Rolling Stones. They
are toy data sets that can be displayed in their entirety on a slide (e.g. to
demonstrate a join).
}
\examples{
analysis_by_group
raw_data
cleaning_log
clean_data
survey
choices
overall_analysis
}
\keyword{datasets}
