% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_data_type.R
\name{find_integer_cols}
\alias{find_integer_cols}
\title{Find all the integer column even if reads as character or factor}
\usage{
find_integer_cols(df)
}
\arguments{
\item{df}{The data set}
}
\description{
Find all the integer column even if reads as character or factor
}
