% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_data_type.R
\name{find_qualitative_cols_from_kobo}
\alias{find_qualitative_cols_from_kobo}
\title{Find all the qualitative using kobo tools}
\usage{
find_qualitative_cols_from_kobo(
  df,
  kobo_survey_sheet,
  additional_cols_to_remove = NULL
)
}
\arguments{
\item{df}{The data set}

\item{kobo_survey_sheet}{The defult (NUll) will create additional table for multiple choice questions}

\item{additional_cols_to_remove}{Any additional columns that you want to include as qualitative}
}
\description{
Find all the qualitative using kobo tools
}
