% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fix_data_type.R
\name{fix_data_type}
\alias{fix_data_type}
\title{Fix data type from the dataset}
\usage{
fix_data_type(
  df,
  remove_all_NA_col = T,
  na.string = c("", "NA"),
  character_cols = NULL
)
}
\arguments{
\item{df}{The data set}

\item{remove_all_NA_col}{A logical parameter. Ture will remove the columns where all values are NA}

\item{na.string}{Na string}

\item{character_cols}{Column that must be character even if it numeric.}
}
\description{
Fix data type from the dataset
}
