% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportive_function.R
\name{left_join_for_list}
\alias{left_join_for_list}
\title{Left join in list}
\usage{
left_join_for_list(list, by)
}
\arguments{
\item{list}{list of dataframes}

\item{by}{primary key that exists in all element}
}
\value{
outcome of left join based on the primary key
}
\description{
Left join in list
}
