% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_write_excel.R
\name{read_sheets}
\alias{read_sheets}
\title{read all the sheets from a Excel file}
\usage{
read_sheets(
  dataset_path,
  remove_all_NA_col = T,
  data_type_fix = T,
  na_strings = c("", "NA", "N/A", " "),
  sheets = "all",
  character_cols = NULL,
  output_as_list = F
)
}
\arguments{
\item{dataset_path}{location of the dataset}

\item{remove_all_NA_col}{A logical parameter. TRUE (default) with remove columns with all NAs}

\item{data_type_fix}{will fix the data type.}

\item{na_strings}{only valid when data_type_fix is true}

\item{sheets}{Sheet to read. Must be the name of the sheet.Default is all, which will read all the available sheets in the excel file.}

\item{character_cols}{name of the columns which must read as characters even if its a non character column.}

\item{output_as_list}{TRUE to have a list with all the excel tabs. FALSE will be making a separate data frame for each tab.}
}
\value{
All of the sheet as individual dataframe
}
\description{
read all the sheets from a Excel file
}
