% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportive_function.R
\name{left_join_for_list}
\alias{left_join_for_list}
\title{Left join all data frames in a list}
\usage{
left_join_for_list(list, by)
}
\arguments{
\item{list}{A list of data frames to be joined.}

\item{by}{Character vector of column name(s) to join by. Must exist in all elements.}
}
\value{
A data frame resulting from left-joining all list elements by the specified key.
}
\description{
Performs a sequential left join on all elements of a list of data frames,
using a common key column present in all elements.
}
\examples{
\dontrun{
df1 <- data.frame(id = 1:3, x = c("A", "B", "C"))
df2 <- data.frame(id = 1:3, y = c(10, 20, 30))
df3 <- data.frame(id = 2:3, z = c(TRUE, FALSE))
list_of_dfs <- list(df1, df2, df3)
left_join_for_list(list_of_dfs, by = "id")
}

}
