% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_analysis.R
\name{survey_analysis}
\alias{survey_analysis}
\title{Batch Collapse Survey Data into tidy long format}
\usage{
survey_analysis(
  df,
  weights = F,
  weight_column = NULL,
  strata,
  vars_to_analyze = NULL,
  disag = NULL,
  na_val,
  sm_sep = "/",
  question_lable = F,
  kobo_path = NULL
)
}
\arguments{
\item{df}{a survey or preferably srvyr object}

\item{weights}{TRUE if you want to consider weights}

\item{weight_column}{column containing weights}

\item{strata}{Strata name}

\item{vars_to_analyze}{columns to collapse}

\item{disag}{the columns to collapse/ subset by(analagous to [[dplyr::group_by]] to [[dplyr::summarise]]) flow}

\item{na_val}{if you want NA replaced by value. By default NA values will be removed prior to aggregation. It is recommended
that you do not adjust this value and deal with na values as a separate step}

\item{sm_sep}{select multiple parent child separator. This is specific for XLSForm data (default = /).}

\item{question_lable}{A logical variable. Select TRUE if label from kobo is necessary in the analysis.
If using read_csv to read in data the separator will most likely be '/' where as if using read.csv it will likely be '.'}

\item{kobo_path}{kobo tool path}
}
\value{
a long format data frame containing the collapsed data.
}
\description{
`survey_analysis` uses the srvyr [srvyr::survey_mean] & survey package [survey::svymean]   methods
to collapse/or aggregate survey data. This function uses `survey_collapse_categorical_long` and `survey_collapse_binary_long`
to perform the batch analysis. This function is extracted from butteR
}
